SECURITY/ERC3643/Oracle/maliciousOracle/requirements.txt
web3
eth_account
python-dotenv